Hooks für OAuth, genutzt von login und signUp
