+++ .gitignore
@@ -30,7 +30,8 @@ yarn-error.*
 .DS_Store
 *.pem
 
-# local env files
+# local env files (Secrets – NIEMALS committen!)
+.env
 .env*.local
 
 # typescript
