+++ app/login.jsx
@@ -5,21 +5,15 @@
  * Apple (iOS): expo-apple-authentication → signInWithIdToken → fetchProfileWithToken → Redirect
  */
 import { View, Text, Pressable, Image, ActivityIndicator, ScrollView, TextInput, TouchableOpacity, StyleSheet, Platform } from "react-native";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { useRouter, Redirect } from 'expo-router';
-import { supabase, fetchProfileWithToken } from "../lib/supabase";
-import { makeRedirectUri } from "expo-auth-session";
-import * as WebBrowser from 'expo-web-browser';
-import * as AppleAuthentication from 'expo-apple-authentication';
-import * as Crypto from 'expo-crypto';
+import { supabase } from "../lib/supabase";
+import { useOAuthHandlers } from "./hooks/useOAuthHandlers";
 import { theme } from "../constants/theme";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { EnvelopeIcon } from "react-native-heroicons/outline";
 import { EyeIcon, EyeSlashIcon } from "react-native-heroicons/outline";
 import { CheckIcon } from "react-native-heroicons/solid";
-
-// WICHTIG: Dies erlaubt dem Browser, die Auth-Session abzuschließen
-WebBrowser.maybeCompleteAuthSession();
 
 export default function Login() {
   const [email, setEmail] = useState("");
@@ -29,19 +23,18 @@ export default function Login() {
   const [errorMsg, setErrorMsg] = useState("");
   const [loading, setLoading] = useState(false);
   const [activeTab, setActiveTab] = useState("signin");
-  const [oauthRedirectTo, setOauthRedirectTo] = useState(null);
-  const [oauthProcessing, setOauthProcessing] = useState(false);
-  const [appleAuthAvailable, setAppleAuthAvailable] = useState(false);
   const router = useRouter();
 
-  // Apple Sign-In Verfügbarkeit prüfen (nur iOS)
-  useEffect(() => {
-    if (Platform.OS === 'ios') {
-      AppleAuthentication.isAvailableAsync().then(setAppleAuthAvailable);
-    }
-  }, []);
+  // OAuth-Logik zentral im Hook (Google + Apple)
+  const {
+    signInWithGoogle,
+    signInWithApple,
+    oauthProcessing,
+    oauthRedirectTo,
+    appleAuthAvailable,
+  } = useOAuthHandlers({ setErrorMsg, mode: 'login', router });
 
-  // Deklarativer Redirect nach OAuth - läuft im Render-Zyklus, zuverlässiger als router.replace
+  // Deklarativer Redirect nach OAuth
   if (oauthRedirectTo) {
     return <Redirect href={oauthRedirectTo} />;
   }
@@ -98,181 +91,6 @@ export default function Login() {
     } else {
       setLoading(false);
       setSubmitted(true);
-    }
-  };
-
-  const handleRegistration = () => {
-    router.push("signup");
-  };
-
-  const signInWithApple = async () => {
-    if (Platform.OS !== 'ios') return;
-
-    try {
-      console.log("[AUTH] Apple Sign-In initiated");
-      setErrorMsg("");
-      setOauthProcessing(true);
-
-      const rawNonce = Crypto.randomUUID();
-      const hashedNonce = await Crypto.digestStringAsync(
-        Crypto.CryptoDigestAlgorithm.SHA256,
-        rawNonce
-      );
-
-      const credential = await AppleAuthentication.signInAsync({
-        requestedScopes: [
-          AppleAuthentication.AppleAuthenticationScope.FULL_NAME,
-          AppleAuthentication.AppleAuthenticationScope.EMAIL,
-        ],
-        nonce: hashedNonce,
-      });
-
-      if (!credential.identityToken) {
-        setOauthProcessing(false);
-        setErrorMsg('Apple-Anmeldung fehlgeschlagen');
-        return;
-      }
-
-      const { data, error } = await supabase.auth.signInWithIdToken({
-        provider: 'apple',
-        token: credential.identityToken,
-        nonce: rawNonce,
-      });
-
-      if (error) {
-        setOauthProcessing(false);
-        setErrorMsg('Apple-Anmeldung fehlgeschlagen: ' + error.message);
-        return;
-      }
-
-      // Full Name nur beim ersten Login – in User-Metadaten speichern
-      if (credential.fullName) {
-        const fullName = [
-          credential.fullName.givenName,
-          credential.fullName.familyName,
-        ].filter(Boolean).join(' ');
-        await supabase.auth.updateUser({
-          data: {
-            full_name: fullName,
-            given_name: credential.fullName.givenName,
-            family_name: credential.fullName.familyName,
-          },
-        });
-      }
-
-      // Profil prüfen und weiterleiten (analog zu Google)
-      const uid = data.session?.user?.id;
-      const accessToken = data.session?.access_token;
-      const profile = accessToken
-        ? await fetchProfileWithToken(accessToken, uid)
-        : (await supabase.from('profiles').select('onboarding_completed').eq('id', uid).maybeSingle()).data;
-      const target = profile?.onboarding_completed ? '/tabs' : '/onboarding';
-      setOauthRedirectTo(target);
-    } catch (err) {
-      setOauthProcessing(false);
-      if (err.code === 'ERR_REQUEST_CANCELED') {
-        setErrorMsg('Apple-Anmeldung abgebrochen');
-      } else {
-        setErrorMsg('Ein unerwarteter Fehler: ' + (err.message || err));
-      }
-    }
-  };
-
-  const signInWithGoogle = async () => {
-    try {
-      console.log("[AUTH] Google Sign-In initiated");
-      setErrorMsg("");
-
-      const redirectUri = makeRedirectUri({
-        scheme: 'n8tly',
-        path: 'auth/callback'
-      });
-
-      console.log("[AUTH] Redirect URI:", redirectUri);
-
-      const { data, error } = await supabase.auth.signInWithOAuth({
-        provider: "google",
-        options: {
-          redirectTo: redirectUri,
-          skipBrowserRedirect: false,
-          queryParams: {
-            access_type: 'offline',
-            prompt: 'consent',
-          }
-        }
-      });
-
-      if (error) {
-        console.error("[AUTH] Google Sign-In Error:", error.message);
-        setErrorMsg('Google-Anmeldung fehlgeschlagen: ' + error.message);
-        return;
-      }
-
-      if (!data?.url) {
-        console.error("[AUTH] No OAuth URL received");
-        setErrorMsg('OAuth-URL konnte nicht erstellt werden');
-        return;
-      }
-
-      console.log("[AUTH] Opening OAuth URL:", data.url);
-
-      const result = await WebBrowser.openAuthSessionAsync(
-        data.url,
-        redirectUri,
-        {
-          showInRecents: true,
-        }
-      );
-
-      console.log("[AUTH] Browser result:", JSON.stringify(result, null, 2));
-
-      if (result.type === 'success' && result.url) {
-        console.log("[AUTH] OAuth successful, URL:", result.url);
-        setOauthProcessing(true); // Sofort Overlay – Login-Screen nicht mehr sichtbar
-
-        const url = new URL(result.url);
-        // Supabase OAuth returns tokens in the hash fragment (#), not query params (?)
-        const hashParams = url.hash ? new URLSearchParams(url.hash.substring(1)) : null;
-        const getParam = (name) => hashParams?.get(name) ?? url.searchParams.get(name);
-        const access_token = getParam('access_token');
-        const refresh_token = getParam('refresh_token');
-
-        if (access_token && refresh_token) {
-          const { data: sessionData, error: sessionError } = await supabase.auth.setSession({
-            access_token,
-            refresh_token,
-          });
-
-          if (sessionError) {
-            console.error("[AUTH] Session Error:", sessionError);
-            setOauthProcessing(false);
-            setErrorMsg('Fehler beim Setzen der Session');
-          } else {
-            // Google OAuth: Profil mit access_token direkt abfragen (siehe fetchProfileWithToken)
-            const uid = sessionData?.session?.user?.id;
-            const accessToken = sessionData?.session?.access_token;
-            const profile = accessToken
-              ? await fetchProfileWithToken(accessToken, uid)
-              : (await supabase.from('profiles').select('onboarding_completed').eq('id', uid).maybeSingle()).data;
-            const target = profile?.onboarding_completed ? '/tabs' : '/onboarding';
-            setOauthRedirectTo(target);
-          }
-        } else {
-          console.error("[AUTH] No tokens in redirect URL");
-          setOauthProcessing(false);
-          setErrorMsg('Anmeldung fehlgeschlagen. Bitte erneut versuchen.');
-        }
-      } else if (result.type === 'cancel') {
-        console.log("[AUTH] User cancelled");
-        setErrorMsg('Google-Anmeldung abgebrochen');
-      } else if (result.type === 'dismiss' || result.type === 'locked') {
-        console.log("[AUTH] Browser dismissed or locked");
-        setErrorMsg('Browser wurde geschlossen');
-      }
-    } catch (err) {
-      console.error("[AUTH] Google Sign-In Exception:", err);
-      setOauthProcessing(false);
-      setErrorMsg('Ein unerwarteter Fehler ist aufgetreten: ' + (err.message || err));
     }
   };
