+++ app/signup.jsx
@@ -1,19 +1,13 @@
 import { View, Text, Pressable, Image, ActivityIndicator, ScrollView, TextInput, TouchableOpacity, Alert, Platform, StyleSheet } from "react-native";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { useRouter, Redirect } from 'expo-router';
-import { supabase, fetchProfileWithToken } from "../lib/supabase";
-import { makeRedirectUri } from "expo-auth-session";
-import * as WebBrowser from 'expo-web-browser';
-import * as AppleAuthentication from 'expo-apple-authentication';
-import * as Crypto from 'expo-crypto';
+import { supabase } from "../lib/supabase";
+import { useOAuthHandlers } from "./hooks/useOAuthHandlers";
 import { theme } from "../constants/theme";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { EnvelopeIcon } from "react-native-heroicons/outline";
 import { EyeIcon, EyeSlashIcon } from "react-native-heroicons/outline";
 import { CheckIcon } from "react-native-heroicons/solid";
-
-// WICHTIG: Dies erlaubt dem Browser, die Auth-Session abzuschließen
-WebBrowser.maybeCompleteAuthSession();
 
 export default function SignUp() {
   const [email, setEmail] = useState("");
@@ -25,17 +19,16 @@ export default function SignUp() {
   const [errorMsg, setErrorMsg] = useState("");
   const [loading, setLoading] = useState(false);
   const [activeTab, setActiveTab] = useState("signup");
-  const [oauthRedirectTo, setOauthRedirectTo] = useState(null);
-  const [oauthProcessing, setOauthProcessing] = useState(false);
-  const [appleAuthAvailable, setAppleAuthAvailable] = useState(false);
   const router = useRouter();
 
-  // Apple Sign-In Verfügbarkeit prüfen (nur iOS)
-  useEffect(() => {
-    if (Platform.OS === 'ios') {
-      AppleAuthentication.isAvailableAsync().then(setAppleAuthAvailable);
-    }
-  }, []);
+  // OAuth-Logik zentral im Hook (Google + Apple)
+  const {
+    signInWithGoogle,
+    signInWithApple,
+    oauthProcessing,
+    oauthRedirectTo,
+    appleAuthAvailable,
+  } = useOAuthHandlers({ setErrorMsg, mode: 'signup', router });
 
   // Deklarativer Redirect nach OAuth
   if (oauthRedirectTo) {
@@ -109,214 +102,6 @@ export default function SignUp() {
 
     Alert.alert('Registrierung erfolgreich! Du kannst dich jetzt anmelden.');
     router.replace('/login');
-  };
-
-  // Nonce-Generator für Apple Sign-In
-  const generateNonce = async () => {
-    const rawNonce = Crypto.randomUUID();
-    const hashedNonce = await Crypto.digestStringAsync(
-      Crypto.CryptoDigestAlgorithm.SHA256,
-      rawNonce
-    );
-    return { rawNonce, hashedNonce };
-  };
-
-  const signInWithGoogle = async () => {
-    try {
-      console.log("[SIGNUP] Google Sign-In initiated");
-      setErrorMsg("");
-      const redirectUri = makeRedirectUri({ scheme: "n8tly", path: "auth/callback" });
-      console.log("[SIGNUP] Redirect URI:", redirectUri);
-
-      const { data, error } = await supabase.auth.signInWithOAuth({
-        provider: "google",
-        options: {
-          redirectTo: redirectUri,
-          skipBrowserRedirect: false,
-          queryParams: {
-            access_type: 'offline',
-            prompt: 'consent',
-          }
-        },
-      });
-
-      if (error) {
-        console.error("[SIGNUP] OAuth initiation error:", error);
-        setErrorMsg("Google-Anmeldung fehlgeschlagen: " + error.message);
-        return;
-      }
-
-      const authUrl = data?.url;
-      if (!authUrl) {
-        setErrorMsg("Keine Auth-URL erhalten");
-        return;
-      }
-
-      console.log("[SIGNUP] Opening browser with URL:", authUrl);
-      const result = await WebBrowser.openAuthSessionAsync(authUrl, redirectUri);
-      console.log("[SIGNUP] Browser result:", result);
-
-      if (result.type !== "success" || !result.url) {
-        console.log("[SIGNUP] OAuth cancelled or failed");
-        return;
-      }
-
-      setOauthProcessing(true);
-
-      const url = new URL(result.url);
-      // Supabase OAuth returns tokens in the hash fragment (#), not query params (?)
-      const hashParams = url.hash ? new URLSearchParams(url.hash.substring(1)) : null;
-      const getParam = (name) => hashParams?.get(name) ?? url.searchParams.get(name);
-      const accessToken = getParam('access_token');
-      const refreshToken = getParam('refresh_token');
-
-      console.log('[SIGNUP] Parsed tokens:', { accessToken: !!accessToken, refreshToken: !!refreshToken });
-
-      if (!accessToken || !refreshToken) {
-        setOauthProcessing(false);
-        setErrorMsg("Keine Tokens erhalten");
-        console.error('[SIGNUP] No tokens in URL:', result.url);
-        return;
-      }
-
-      const { data: sessionData, error: sessionError } = await supabase.auth.setSession({
-        access_token: accessToken,
-        refresh_token: refreshToken,
-      });
-
-      if (sessionError || !sessionData?.session) {
-        setOauthProcessing(false);
-        setErrorMsg("Session konnte nicht gesetzt werden");
-        return;
-      }
-
-      const session = sessionData.session;
-      const uid = session.user.id;
-
-      // PRÜFUNG: Ist der User neu? (Registrierung) oder besteht bereits ein Account? (Login)
-      const userCreatedAt = new Date(session.user.created_at);
-      const now = new Date();
-      const ageInSeconds = (now - userCreatedAt) / 1000;
-
-      if (ageInSeconds > 10) {
-        // User existiert bereits - Registrierung abbrechen
-        console.log('[SIGNUP] User already exists, created at:', session.user.created_at);
-        await supabase.auth.signOut();
-        setOauthProcessing(false);
-        Alert.alert(
-          'Account existiert bereits',
-          'Dieser Account existiert bereits. Du wirst zum Login weitergeleitet.',
-          [{ text: 'OK', onPress: () => router.replace('/login') }]
-        );
-        return;
-      }
-
-      // Neuer User - weiter zum Onboarding
-      console.log('[SIGNUP] New user created:', uid);
-      const profile = await fetchProfileWithToken(session.access_token, uid);
-      console.log('[SIGNUP] Profile:', profile);
-
-      const onboardingComplete = profile?.onboarding_completed === true;
-      if (onboardingComplete) {
-        setOauthRedirectTo('/tabs');
-      } else {
-        setOauthRedirectTo('/onboarding');
-      }
-    } catch (err) {
-      console.error("[SIGNUP] Google Sign-In error:", err);
-      setOauthProcessing(false);
-      setErrorMsg("Ein Fehler ist aufgetreten: " + err.message);
-    }
-  };
-
-  const signInWithApple = async () => {
-    if (Platform.OS !== 'ios') return;
-
-    try {
-      console.log("[SIGNUP] Apple Sign-In initiated");
-      setErrorMsg("");
-      setOauthProcessing(true);
-
-      const { rawNonce, hashedNonce } = await generateNonce();
-
-      const credential = await AppleAuthentication.signInAsync({
-        requestedScopes: [
-          AppleAuthentication.AppleAuthenticationScope.FULL_NAME,
-          AppleAuthentication.AppleAuthenticationScope.EMAIL,
-        ],
-        nonce: hashedNonce,
-      });
-
-      if (!credential.identityToken) {
-        setOauthProcessing(false);
-        setErrorMsg('Apple-Anmeldung fehlgeschlagen');
-        return;
-      }
-
-      const { data, error } = await supabase.auth.signInWithIdToken({
-        provider: 'apple',
-        token: credential.identityToken,
-        nonce: rawNonce,
-      });
-
-      if (error) {
-        setOauthProcessing(false);
-        setErrorMsg('Apple-Anmeldung fehlgeschlagen: ' + error.message);
-        return;
-      }
-
-      const session = data.session;
-      const uid = session.user.id;
-
-      // PRÜFUNG: Ist der User neu? (Registrierung) oder besteht bereits ein Account? (Login)
-      const userCreatedAt = new Date(session.user.created_at);
-      const now = new Date();
-      const ageInSeconds = (now - userCreatedAt) / 1000;
-
-      if (ageInSeconds > 10) {
-        // User existiert bereits - Registrierung abbrechen
-        console.log('[SIGNUP] User already exists, created at:', session.user.created_at);
-        await supabase.auth.signOut();
-        setOauthProcessing(false);
-        Alert.alert(
-          'Account existiert bereits',
-          'Dieser Account existiert bereits. Du wirst zum Login weitergeleitet.',
-          [{ text: 'OK', onPress: () => router.replace('/login') }]
-        );
-        return;
-      }
-
-      // Neuer User - fullName speichern (falls vorhanden)
-      if (credential.fullName) {
-        const fullName = [credential.fullName.givenName, credential.fullName.familyName]
-          .filter(Boolean)
-          .join(' ');
-        if (fullName) {
-          await supabase.auth.updateUser({
-            data: { full_name: fullName }
-          });
-        }
-      }
-
-      console.log('[SIGNUP] New user created:', uid);
-      const profile = await fetchProfileWithToken(session.access_token, uid);
-      console.log('[SIGNUP] Profile:', profile);
-
-      const onboardingComplete = profile?.onboarding_completed === true;
-      if (onboardingComplete) {
-        setOauthRedirectTo('/tabs');
-      } else {
-        setOauthRedirectTo('/onboarding');
-      }
-    } catch (err) {
-      console.error("[SIGNUP] Apple Sign-In error:", err);
-      setOauthProcessing(false);
-      if (err.code === 'ERR_REQUEST_CANCELED') {
-        console.log('[SIGNUP] User cancelled Apple Sign-In');
-      } else {
-        setErrorMsg("Ein Fehler ist aufgetreten: " + err.message);
-      }
-    }
   };
 
   return (
